src/requirements.txt
requests==2.22.0
APScheduler==3.9.1
pymongo==4.3.3
python-dotenv==0.21.0